+++ src/io/segment/android/Analytics.java
@@ -1000,7 +1000,7 @@ public class Analytics {
 			context = new Context();
 		
 		Alias alias = new Alias(from, to, timestamp, context);
-
+		
 		enqueue(alias);
 		
 		providerManager.alias(alias);