+++ history.md
@@ -1,8 +1,9 @@
 0.2.0 / 2013-05-08
 =================
 * Adding ability to request integration settings from Segment.io
-* Adding bundled providers to the library
 * Adding Amplitude, BugSnag, Countly, Crittercism, Flurry, Google Analytics, Localytics, and Mixpanel bundled providers and associated libraries
+* Adding context.providers support for identifies, aliases, and tracks
+* Added stopwatch timing for bundled provider operations
 
 0.1.1 / 2013-04-23
 =================